pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docx2txt==0.8",
    "faiss-cpu==1.7.4",
    "gradio==4.13.0",
    "langchain>=0.1.7",
    "langchain-community==0.0.20",
    "langchain-groq>=0.0.1",
    "ollama==0.1.2",
    "pandas==2.0.3",
    "plotly>=6.3.0",
    "pypdf==3.17.0",
    "python-docx==1.1.0",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "sentence-transformers==2.2.2",
    "unstructured==0.13.4",
]
